requirements.txt
torch
moviepy
opencv-python
pillow
numpy
transformers
diffusers
safetensors
peft
decord